+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_constants.hpp
@@ -1,0 +1,12 @@
+
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+#pragma once
+
+namespace Azure { namespace Storage { namespace Files { namespace DataLake { namespace Details {
+  // Error codes:
+  constexpr static const char* DataLakeFilesystemNotFound = "FilesystemNotFound";
+  constexpr static const char* DataLakePathNotFound = "PathNotFound";
+
+}}}}} // namespace Azure::Storage::Files::DataLake::Details