+++ sdk/storage/azure-storage-files-shares/test/share_file_client_test.cpp
@@ -69,6 +69,37 @@ namespace Azure { namespace Storage { namespace Test {
         EXPECT_NO_THROW(client.Create(1024));
       }
     }
+    {
+      // DeleteIfExists.
+      {
+        auto client = m_shareClient->GetRootShareDirectoryClient().GetShareFileClient(
+            LowercaseRandomString());
+        EXPECT_NO_THROW(client.Create(1024));
+        EXPECT_NO_THROW(client.Delete());
+        EXPECT_NO_THROW(client.DeleteIfExists());
+      }
+      {
+        auto client = m_shareClient->GetRootShareDirectoryClient().GetShareFileClient(
+            LowercaseRandomString());
+        auto deleteResult = client.DeleteIfExists();
+        EXPECT_FALSE(deleteResult->Deleted);
+      }
+      {
+        auto shareClient = Files::Shares::ShareClient::CreateFromConnectionString(
+            StandardStorageConnectionString(), LowercaseRandomString());
+        auto client = m_shareClient->GetRootShareDirectoryClient().GetShareFileClient(
+            LowercaseRandomString());
+        auto deleteResult = client.DeleteIfExists();
+        EXPECT_FALSE(deleteResult->Deleted);
+      }
+      {
+        auto client = m_shareClient->GetRootShareDirectoryClient()
+                          .GetSubShareDirectoryClient(LowercaseRandomString())
+                          .GetShareFileClient(LowercaseRandomString());
+        auto deleteResult = client.DeleteIfExists();
+        EXPECT_FALSE(deleteResult->Deleted);
+      }
+    }
   }
 
   TEST_F(FileShareFileClientTest, FileMetadata)
@@ -159,7 +190,8 @@ namespace Azure { namespace Storage { namespace Test {
       Files::Shares::CreateFileOptions options3;
       options3.SmbProperties.PermissionKey = result1;
       std::string permissionKey;
-      EXPECT_NO_THROW(permissionKey = client3.Create(1024, options3)->FilePermissionKey);
+      EXPECT_NO_THROW(
+          permissionKey = client3.Create(1024, options3)->SmbProperties.PermissionKey.GetValue());
       auto result3 = client3.GetProperties()->FilePermissionKey;
       EXPECT_EQ(permissionKey, result3);
     }